integrated the avatarcolor util into component
